microcloud/service: Update package usages

+++ microcloud/service/service_handler.go
@@ -19,16 +19,16 @@ import (
 
 const (
 	// OVNPort is the efault MicroOVN port.
-	OVNPort int = 6443
+	OVNPort int64 = 6443
 
 	// CephPort is the efault MicroCeph port.
-	CephPort int = 7443
+	CephPort int64 = 7443
 
 	// LXDPort is the efault LXD port.
-	LXDPort int = 8443
+	LXDPort int64 = 8443
 
 	// CloudPort is the efault MicroCloud port.
-	CloudPort int = 9443
+	CloudPort int64 = 9443
 )
 
 // Handler holds a set of services and an mdns server for communication between them.
@@ -38,7 +38,7 @@ type Handler struct {
 	Services map[types.ServiceType]Service
 	Name     string
 	Address  string
-	Port     int
+	Port     int64
 
 	AuthSecret string
 }
